requirements.txt
google-cloud-storage
google-cloud-aiplatform
tensorflow
pandas
numpy
Pillow